# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Postqron API
  # 1.1.0: two additive routes for consent to the legal documents (R46), and an
  # optional `language` on registration. Nothing existing changed shape, so a
  # minor bump. The versioning and deprecation *policy* is R52, issue #466: when
  # it lands it decides the rule, and this number follows it.
  version: 1.1.0
  summary: Manage, sync and monitor cron jobs (R8, R9, R51).
  description: |
    The contract of the Postqron public API. It describes **what the service
    actually does**: paths, bodies, error codes and limits are compared against
    the code on every `make ci` run — see "What keeps this document honest".

    ## What keeps this document honest

    A hand-written contract drifts from the code within a month, and from that
    moment it is worse than having none: whoever reads it builds a client on a
    false promise, and the defect shows up at their end. Here drift is prevented
    by an automated check
    (`services/api/internal/httpapi/contract_test.go`) that compares the
    document against the code and fails CI when they disagree:

    - **routes** — the `(method, path)` list comes from the real router, not
      from a second hand-written copy;
    - **permissions** — `x-api-key-scope` is compared against the scope the code
      enforces, and no operation may list `apiKey` among its authentication
      methods unless the code registers it behind a scope;
...
      exposing something;
...

...

    ## Contract version
...
servers:
...
security:
  - sessionCookie: []
  - sessionBearer: []
paths:
...
  /keys:
    get:
      tags: [api keys]
      operationId: listApiKeys
      summary: List API keys (R9)
      description: |
        **Session only.** An API key cannot list, create or revoke keys: if it
        could, a read-only key would be one request away from issuing a write
        key, and the whole scope system would become a formality. The practical
        consequence — automation cannot rotate its own keys — is intended:
        rotation is something a person decides.
      parameters:
        - name: include_revoked
          in: query
          description: |
            Include revoked keys, which stay in the table as a historical trace:
            whoever investigates an incident wants to see them.
          schema: { type: boolean }
      responses:
        "200":
          description: The user's keys. **None of them carries its own value.**
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyList" }
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [api keys]
      operationId: createApiKey
      summary: Issue an API key (R9)
      description: |
        **This is the only response in the whole API that contains a secret**, and
        the only moment that value exists: afterwards it is not recoverable
        anywhere, by anyone, not even by an administrator. There is no `GET
        /keys/{id}` and no way to read it back — if it is lost, revoke it and
        create another.

        The response carries `Cache-Control: no-store`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiKeyInput" }
      responses:
        "201":
          description: Key created. `secret` is visible **here and now only**.
          headers:
            Cache-Control:
              description: Always `no-store`, because the body carries a secret.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          description: "`too_many_api_keys` — revoke one before creating another."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /secrets:
    get:
      tags: [secrets]
      operationId: listSecrets
      summary: List workspace secrets (R42)
      description: |
        **No route in this API returns the value of a secret**, and it is not a
        forgotten route: there is no `GET /secrets/{id}`, no `reveal` parameter,
        no administrative variant. There is not even a `last_four` preview: that
        makes sense for a known provider's key, recognisable by its tail, not for
        an arbitrary customer value that might *be* four meaningful characters.

        **Session only.** An API key able to replace a secret's value would
        redirect the credentials with which the user's jobs authenticate to their
        own services, and no log of theirs would show a changed value.
      parameters:
        - name: include_revoked
          in: query
          description: Include revoked secrets, kept as a trace for whoever investigates a failed execution.
          schema: { type: boolean }
      responses:
        "200":
          description: The workspace secrets, **without their values**.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretList" }
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [secrets]
      operationId: createSecret
      summary: Store a secret (R42)
      description: |
        The value **does not come back even on creation**: unlike an API key,
        which we generate, this one the user already has — they just pasted it.

        The response carries `Cache-Control: no-store` even though it contains no
        value: the *request* does, and an intermediary caching the exchange would
        keep a copy.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretInput" }
      responses:
        "201":
          description: Secret stored. **The response does not contain the value.**
          headers:
            Cache-Control:
              description: Always `no-store`, because of the request body.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Secret" }
        "400":
          description: |
            `validation_failed` or `invalid_request`. On this route the message
            **does not quote** the text that could not be read: on a request
            carrying a secret, that text may be the secret.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          description: |
            `secret_name_taken` — a live secret with that name already exists:
            update its value instead of creating another. `too_many_secrets`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /account/deletion:
    get:
      tags: [account]
      operationId: getAccountDeletion
      summary: State of the deletion request (R45)
      description: |
        **The resource is the deletion request, not the account**: it is what gets
        created, inspected and withdrawn, and it has a state that lasts for the
        grace period. A `DELETE /account` would have suggested an instantaneous
        operation, which is exactly what this is not.

        It also answers before the user asks: `grace_hours` says how long they
        will have to change their mind, and `subscription` says whether a paid
        subscription would stay alive — so the confirmation screen can state it
        instead of discovering it from a rejection.

        **Session only**, the strictest choice available and the one with the most
        at stake: an API key able to request deletion would make a service
        credential forgotten in a config file enough to destroy someone's work.
      responses:
        "200":
          description: The state of the deletion request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionStatus" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [account]
      operationId: requestAccountDeletion
      summary: Request account deletion (R45)
      description: |
        Opens the grace window, stops executions and revokes credentials. It
        answers **`202` and not `200`**: the request is accepted and what it
        promises — removal of the data — happens after the grace period. A `200`
        would say it is done.

        The response counts what was stopped, because the privacy policy promises
        executions stop and credentials are revoked *immediately*: an answer that
        only said "fine" would leave the user to verify it elsewhere.

        The **password is required** as confirmation (R45). If a paid subscription
        is alive, `subscription_acknowledged` is required too: closing the account
        does not cancel the Paddle subscription, and the already-paid period is
        not refunded — a fact the user must be shown before, not after.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeletionRequestInput" }
      responses:
        "202":
          description: Deletion requested. Executions stopped, credentials revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionReceipt" }
        "400":
          description: |
            `validation_failed` — the password field is missing, with `details`
            pointing at it; `invalid_request` — unreadable body or unknown field.
            The message does not quote the unreadable text: this body carries a
            password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: "`unauthenticated`, or `invalid_credentials` when the confirmation password is wrong."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "409":
          description: |
            `deletion_already_requested` — a deletion is already in progress and
            can be withdrawn until the grace period expires;
            `subscription_active` — a paid subscription is alive and
            `subscription_acknowledged` was not sent. The latter carries `plan` in
            the body, so the client can say *what* stays alive without a second
            request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [account]
      operationId: cancelAccountDeletion
      summary: Withdraw the deletion request (R45)
      description: |
        Available until the grace period expires. Jobs come back; **credentials do
        not**, and their absence is the fact: revocation emptied the encrypted
        material, so there is nothing to restore. Whoever comes back finds their
        jobs and their data, and issues new keys.
      responses:
        "200":
          description: Request withdrawn, jobs resumed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionRestored" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          description: "`deletion_not_requested` — there is nothing to withdraw."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # -------------------------------------------------------------------- legal
  /legal/consents:
    get:
      tags: [legal]
      operationId: getLegalConsents
      summary: Consents given to the legal documents (R46)
      description: |
        Answers three questions at once: what the user has accepted, what is
        missing, and what is about to change.

        **`accepted` keeps superseded versions.** They are not history for its
        own sake: if a clause is ever disputed, the question is what the user had
        in front of them when they accepted, and the answer must be
        reconstructible without interpretation — which version, in which
        language, on which day.

        **The four documents are not on the same version**, and never will be
        again: today the Terms are at 1.2.0, the Privacy Policy at 1.1.0, the
        other two at 1.0.0. A single version number for the set would already be
        false.

        `upcoming` is the half of Terms §9 that can be shown: a version that has
        been announced and does not bind yet. Whoever does not want to accept it
        can close the account before it takes effect (`POST /account/deletion`).

        **Session only**: accepting a contract is an act of the person. An API
        key able to do it would let a service credential forgotten in a config
        file bind someone to a document they never read.
      parameters:
        - name: language
          in: query
          required: false
          schema: { $ref: "#/components/schemas/Language" }
          description: |
            The language the texts to accept will be shown in. It does **not**
            affect consents already given, which carry the language they had —
            that is the whole point of recording it. Absent means English.
      responses:
        "200":
          description: The user's consents.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalConsents" }
        "400":
          description: "`validation_failed` — the `language` parameter is not one of the five."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [legal]
      operationId: acceptLegalDocuments
      summary: Accept the documents in force (R46)
      description: |
        Records the proof and returns the updated state.

        **The client must declare which version it showed.** Between the page
        loading and the button being pressed a new version can take effect: if
        the service recorded "whatever is in force now", the proof would say
        something false precisely in the case that matters — the one where the
        text changed. Declaring it turns the mismatch into a `409` and the client
        shows the right text again.

        **`200`, not `201`.** The resource is the user's consent state, which
        already exists. Re-accepting a version already accepted is not an error,
        creates nothing, and **does not move the date of the first acceptance** —
        that date is the instant the user became bound, and a double form
        submission must not be able to shift it.

        This route is not where consent is born for a new account: creating an
        account **is** the acceptance (Terms, opening paragraph), and the proof is
        written in the same transaction as the account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LegalAcceptInput" }
      responses:
        "200":
          description: Consents recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalConsents" }
        "400":
          description: |
            `validation_failed` — empty list, unknown document, missing version,
            unsupported language, or the same document twice, with `details`
            pointing at the field; `invalid_request` — unreadable body or unknown
            field.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          description: |
            `legal_version_not_in_force` — the version accepted is not the one in
            force. The message carries both numbers, because the difference
            between "you accepted a superseded version" and "you accepted one
            that is not in force yet" is the difference between reloading and
            waiting.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------------- webhooks
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pq_session
      description: |
        Browser session (R14). `HttpOnly`: the token is out of reach of
        JavaScript, so an XSS on the dashboard cannot carry it away.

    sessionBearer:
      type: http
      scheme: bearer
      description: |
        The same session token, for clients without a cookie jar. When both
        arrive, the cookie wins — it is the session the user expects to be using.

    apiKey:
      type: http
      scheme: bearer
      description: |
        Scoped API key (R9): `Authorization: Bearer pq_live_...`. The prefix is
        what tells it apart from a session token without having to try both.

        The scope each operation requires is declared in `x-api-key-scope`, and it
        is compared against the code on every CI run. Write scopes do **not**
        imply read scopes: otherwise someone looking at a key's scopes would see
        `jobs:write` and have to remember that it covers reading too.

        Operations that do not declare this scheme **do not accept a key**, ever:
        they are the ones that issue credentials, commit to a payment, or destroy
        the account.

    operatorToken:
      type: http
      scheme: bearer
      description: |
        Operator token (`POSTQRON_METRICS_TOKEN`), for the metrics and for the
        detail of the readiness answer. **It is not a user credential**: this is
        information about the *service*, and tying it to an API key would mean
        that watching the engine requires some customer's account, and that
        revoking that key turns monitoring off.
...
  responses:
...
    Unauthenticated:
      description: |
        `unauthenticated` — session missing or expired; or `invalid_api_key` —
        key missing, revoked or expired (R9). Two distinct codes because the
        remedy differs: whoever gets the second does not need to sign in again,
        they need to check the key. The message does not distinguish
        "nonexistent" from "revoked" from "expired": that difference would turn
        the API into an oracle about a key found lying around.
      headers:
        WWW-Authenticate:
          description: '`Bearer error="invalid_token"` when the credential was an API key.'
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationFailed:
      description: |
        `validation_failed` with `details` per field, or `invalid_request` if the
        body is unreadable or carries an unknown field. A misspelled `passwrod`
        would otherwise be an empty password, with no way for the client to
        notice.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    RateLimited:
      description: |
        `rate_limited` — a **technical ceiling** on requests (R10). It is a defence
        of the service, identical on every plan: the body has **no `plan` and no
        `limit`**, because no plan grants more of it and suggesting an upgrade
        would be a commercial lie.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServiceLimitEnvelope" }
...
    InternalError:
      description: |
        `internal_error`. The message is generic by construction: the detail might
        contain information about the database, and it belongs in the service log.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ApiKeyScope:
      type: string
      enum: ["jobs:read", "jobs:write", "executions:read", "executions:trigger"]
...
    ErrorEnvelope:
      type: object
      description: The envelope of every error response.
      properties:
        error: { $ref: "#/components/schemas/Error" }
      required: [error]
...
    ApiKeyList:
      type: object
      properties:
        keys:
          type: array
          items: { $ref: "#/components/schemas/ApiKey" }
      required: [keys]
...
    ApiKeyInput:
      type: object
      properties:
        name: { type: string }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/ApiKeyScope" }
        expires_at:
          type: [string, "null"]
          format: date-time
          description: Absent means "never expires".
      additionalProperties: false

    # ----------------------------------------------------------------- secrets
...